[add] 주문/결제 Controller, Service, Mapper 생성

+++ src/main/webapp/WEB-INF/views/cart/cartList.jsp
@@ -34,33 +34,41 @@
 					</div>
 					
 					<div class="listing__details__comment cartList">
-						<c:forEach items="${cartProductList }" var="cartProduct">
-							
-	                            <div class="listing__details__comment__item cartProduct_item">
-		                            	<!-- 체크박스 -->
-		                            	<div class="listing__details__comment__item__pic checkboxPronum">
-		                                    <input type="checkbox" 
-		                                    	data-pronum="<c:out value="${cartProduct.pronum}" />"
-		                                    	data-price="<c:out value="${cartProduct.price}" />">
-		                                </div>
-		                            	<!-- 썸네일 -->
-		                                <div class="listing__details__comment__item__pic">
-		                                    <img src="/resources/img/listing/details/comment.png" alt="">
-		                                </div>
-		                                <div class="listing__details__comment__item__text productData">
-		                                 	<!-- 삭제버튼 -->
-		                                    <div class="listing__details__comment__item__exit exitButton">
-		                                       	<a href="${cartProduct.pronum}"><i class="fa fa-times"></i></a>
-		                                    </div>
-		                                    <!-- 상품정보 -->
-		                                   	<b>상품명 : <c:out value="${cartProduct.title}" /></b>
-		                                   	<p>판매자 : <c:out value="${cartProduct.id}" /></p>
-		                                    <p>상품설명 : <c:out value="${cartProduct.prodsc}" /></p>
-		                                 	가격 : <c:out value="${cartProduct.price}" />원
-		                               	</div>
-	                           	</div>
-	                        
-						</c:forEach>
+						
+						<form action="/order/order.fr" id="checkForm" method="post">
+							<!-- 아이디는 세션에서 가져온다 -->
+							<input type="hidden" name="id" id="id" value="user3">
+							<input type="hidden" name="price" id="price" value="">
+					
+							<c:forEach items="${cartProductList }" var="cartProduct">
+								
+		                            <div class="listing__details__comment__item cartProduct_item">
+			                            	<!-- 체크박스 -->
+			                            	<div class="listing__details__comment__item__pic checkboxPronum">
+			                                    <input type="checkbox" name="checkPronum"
+			                                    	value="<c:out value="${cartProduct.pronum}" />"
+			                                    	data-price="<c:out value="${cartProduct.price}" />">
+			                                </div>
+			                            	<!-- 썸네일 -->
+			                                <div class="listing__details__comment__item__pic">
+			                                    <img src="/resources/img/listing/details/comment.png" alt="">
+			                                </div>
+			                                <div class="listing__details__comment__item__text productData">
+			                                 	<!-- 삭제버튼 -->
+			                                    <div class="listing__details__comment__item__exit exitButton">
+			                                       	<a href="${cartProduct.pronum}"><i class="fa fa-times"></i></a>
+			                                    </div>
+			                                    <!-- 상품정보 -->
+			                                   	<b>상품명 : <c:out value="${cartProduct.title}" /></b>
+			                                   	<p>판매자 : <c:out value="${cartProduct.id}" /></p>
+			                                    <p>상품설명 : <c:out value="${cartProduct.prodsc}" /></p>
+			                                 	가격 : <c:out value="${cartProduct.price}" />원
+			                               	</div>
+		                           	</div>
+		                        
+							</c:forEach>
+						
+						</form>
 					</div>
 					
 				</div>
@@ -71,7 +79,7 @@
 					<div class="blog__sidebar__recent orderPrice">
 						<h5>선택상품</h5>
 						결제예정금액 : <b>0</b>원 <br><br>
-						<button class="site-btn buttonTest">주문하기</button>
+						<button class="site-btn buttonTest">결제하기</button>
 					</div>
 					
 				</div>
@@ -89,10 +97,12 @@
 		var cartProductList = '<c:out value="${cartProductList.isEmpty() }" />';
 // 		console.log(typeof(cartProductList));
 		
-/* 		// 즉시 실행 함수
+ 		// 즉시 실행 함수
 		(function(){
-			
-		})();	// end function */
+			$("input[type='checkbox']").each(function(){
+				$(this).prop("checked", false);
+			});
+		})();	// end function 
 		
 		checkCartList(cartProductList);
 		
@@ -102,20 +112,19 @@
 				$(".cartList").html("<h5>장바구니에 담긴 상품이 없습니다.</h5>");
 			}
 		} 
-
-		// 주문하기 버튼을 누르면 주문 페이지에 상품번호와 아이디를 전달(List<idPronumVO>)
+		
+		// 주문하기 버튼을 누르면 주문 페이지에 상품번호와 아이디, 총 금액을 전달
 		$(".buttonTest").click(function(){
-			console.log(typeof(Number($(".orderPrice b").html())));
+			//console.log(typeof(Number($(".orderPrice b").html())));
 			if($(".orderPrice b").html() == "0"){
 				alert("상품을 한 개 이상 선택해 주세요.");
+				return;
 			}
-				
-			$(".checkboxPronum input[type='checkbox']:checked").each(function(){
-
-				console.log($(this).data("pronum"));
-			});
+			
+			$("input[id='price']").attr("value", price);
+			$("#checkForm").submit();
 		});
-		
+
 		// 총 금액
 		var price = 0;
 		$(".checkboxPronum input[type='checkbox']").change(function(){
@@ -170,18 +179,23 @@
 		
 		// 모든 상품 삭제
 		$("#allDelete").click(function(e){
-			e.preventDefault();
-			var id = $(this).attr("href");
-			$(".cartList").html("<h5>장바구니에 담긴 상품이 없습니다.</h5>");
 			
-			$.ajax({
-				url : '/cart/deleteAllProcess.fr/' + id,
-				dataType : 'text',
-				type : 'DELETE',
-				success : function(result){
-					alert(result);
-				}
-			});	 // end ajax
+			e.preventDefault();
+			if(confirm("장바구니를 비우시겠습니까?") == true){
+				var id = $(this).attr("href");
+				$(".cartList").html("<h5>장바구니에 담긴 상품이 없습니다.</h5>");
+				
+				$.ajax({
+					url : '/cart/deleteAllProcess.fr/' + id,
+					dataType : 'text',
+					type : 'DELETE',
+					success : function(result){
+						alert(result);
+					}
+				});	 // end ajax
+			} else {
+				return;
+			}
 		});	
 		
 	});
